telegram_bot_randomlab/requirements.txt
python-telegram-bot
python-dotenv
pytest
pytest-asyncio
pytest-html
